Added tests for checking if client has disconnected and removing from list. Server is using disconnect function to remove clients when it needs to.
client.py now runs in loop and server.py will continue to send new sentences after client.py is done. When there is no verb in sentence from server, the client shuts down and restarts in a new shell.

Added a bunch of prints to console on both server and client to make the user experience much better and more informative.

Next: Client does not close thread on server when closing window. Need to make the server know client has disconnected when this is happening too.