Now a functioning chat room!
Added a function for handling input from server that runs in a thread.

The client now get asked for a nickname which has to be the name of the bot the client is going to be. This might get changed later for something like a list or something.
The client response to its version of a bot with a "match" switch.

The server is the one taking in all responses and addinhg them on screen. Might add so client windows also can see the others response.

The server seems to handle disconnects and such much better now and has not crashed yet during my small testing.

Next time: add input to script, ip, port and bot.